feat(web-user-home): add user profile home workflows
Add user profile update, theme update, profile detail, published video, collection, and home-page series workflows.

Refresh token information when profile identity fields change and expose public user-home data for profile pages.

+++ nilo-common/src/main/java/com/neon/nilocommon/entity/po/UserInfo.java
@@ -43,5 +43,5 @@ public class UserInfo
 
     private Integer currentCoin;
 
-    private Integer theme;
+    private Short theme;
 }
